+++ sdmx/writer/__init__.py
@@ -1,11 +1,9 @@
 from .csv import to_csv
 from .pandas import to_pandas
-from .xml import install_schemas, to_xml, validate_xml
+from .xml import to_xml
 
 __all__ = [
-    "install_schemas",
     "to_csv",
     "to_pandas",
     "to_xml",
-    "validate_xml",
 ]
